% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_methods.R
\name{construct_query}
\alias{construct_query}
\title{create the url extension for an iCite API call}
\usage{
construct_query(
  query_type = "pmid",
  pmids = NA,
  year = NA,
  offset = NA,
  limit = NA
)
}
\arguments{
\item{query_type}{character. The type of query being constructed. Limited to
'pmid' or 'search'. If 'pmid', then you must provide a vector of pmids.}

\item{pmids}{character or numeric. A vector of pmids to be queried. If
`query_type` is set to 'pmid', this is the only argument used. All others
will be ignored.}

\item{year}{character or numeric. The year to be searched.}

\item{offset}{character. The lowest PMID to be returned.}

\item{limit}{character. Maximum number of entries to return.}
}
\description{
\code{construct_query} constructs the url extension for an iCite API call
}
\keyword{internal}
