% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_methods.R
\name{icite_api}
\alias{icite_api}
\title{Query the iCite api}
\usage{
icite_api(pmid)
}
\arguments{
\item{pmid}{character. The pubmed ID to be queried}
}
\value{
If the call runs without error, the output is a simple S3 object
}
\description{
\code{icite_api} returns the parsed results of a single call to the iCite api
}
\examples{
dat <- icite_api('27599104')
print(dat)
}
