% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iCiteR.R
\docType{package}
\name{iCiteR}
\alias{iCiteR}
\alias{iCiteR-package}
\title{iCiteR: A package for retrieving citation data from the NIH's iCite API}
\description{
The iCiteR package provides one primary function: get_metrics()
Users wanting to explore the output of the iCite API in more detail may
also be interested in the icite_api() function, which provides slightly
more informative error messages, but can only handle one PMID at a time.
}
