% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metrics.R
\name{page_no_limit}
\alias{page_no_limit}
\title{Pages through iCite results until there is no new data}
\usage{
page_no_limit(year = year, offset = offset)
}
\arguments{
\item{year}{character. The year whose data you would like}

\item{offset}{character. The minimum PMID you would like to have
returned}
}
\value{
All values from the icite_api reformatted into a dataframe.
see \url{https://icite.od.nih.gov/api} for details.
}
\description{
\code{page_no_limit} aggregates results from iCite until there is no new
data returned
}
\keyword{internal}
