% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iCiteR.R
\docType{package}
\name{iCiteR}
\alias{iCiteR}
\alias{iCiteR-package}
\title{iCiteR: A package for retrieving citation data from the NIH's iCite API}
\description{
The iCiteR package provides a set of functions to facilitate interaction
with the iCite API. The primary functions are get_metrics() and
search_metrics(). The former is used when the user has a known set of PMIDs
and wishes to obtain information about those IDs. The latter is used for
getting data fromm the iCite API, given a set of search parameters.
}
