% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_methods.R
\name{print.icite_api}
\alias{print.icite_api}
\title{A simple print method for the icite_api function}
\usage{
\method{print}{icite_api}(x, ...)
}
\arguments{
\item{x}{S3. The S3 object yielded from \code{icite_api}}

\item{...}{Additional arguments that would be passed to \code{print}}
}
\value{
Prints details of the request, and the content returned from the \code{icite_api} function.
}
\description{
\code{print.icite_api} prints the data returned from the icite_api function
}
\examples{
dat <- icite_api('27599104')
print(dat)
}
