% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metrics.R
\name{to_dataframe}
\alias{to_dataframe}
\title{Converts icite_api information into a dataframe}
\usage{
to_dataframe(info, error = FALSE)
}
\arguments{
\item{info}{S3. The S3 object returned from \code{icite_api}}

\item{error}{logical. Indicates whether the \code{icite_api} call resulted in an error.}
}
\value{
All values from the icite_api reformatted into a dataframe.
see \url{https://icite.od.nih.gov/api} for details.
}
\description{
\code{to_dataframe} converts the parsed json from icite_api into a dataframe
}
\keyword{internal}
